Commit from IntelliJ-12
